Fix Issue #20: Numerical errors with high propensities at long times
Fix the issue by representing time as double rather than float. In
addition, if the time between events dt is more than 12 orders of
magnitude (i.e. 10^12) times smaller than the time t, throw an
exception. This ensures that there are at least 13 bits available to
represent dt, since the mantissa in a double has 53 bits, and
log2(10^12) (the maxiumum ratio between t and dt) is approximately 40
bits.
https://docs.oracle.com/javase/specs/jls/se7/html/jls-4.html#jls-4.2.3

+++ src/java/org/demonsoft/spatialkappa/model/SimulationState.java
@@ -5,9 +5,9 @@ import java.util.Map;
 
 public interface SimulationState {
 
-    float getTime();
-    float getElapsedTime();
-    float getMaximumTime();
+    double getTime();
+    double getElapsedTime();
+    double getMaximumTime();
     int getEventCount();
     int getMaximumEventCount();
